comments stuff

+++ Repositories/ForumPostRepository.cs
@@ -1,4 +1,5 @@
 ﻿
+using Microsoft.EntityFrameworkCore;
 using TheAgoraAPI.Models;
 
 namespace TheAgoraAPI.Repositories
@@ -8,11 +9,13 @@ namespace TheAgoraAPI.Repositories
 
         private readonly TheAgoraDbContext dbContext;
         private readonly ILikeRepository likeRepository;
+        private readonly ICommentRepository commentRepository;
 
-        public ForumPostRepository(TheAgoraDbContext dbContext,ILikeRepository likeRepository)
+        public ForumPostRepository(TheAgoraDbContext dbContext, ILikeRepository likeRepository, ICommentRepository commentRepository)
         {
             this.dbContext = dbContext;
             this.likeRepository = likeRepository ?? throw new ArgumentNullException(nameof(likeRepository));
+            this.commentRepository = commentRepository ?? throw new ArgumentNullException(nameof(commentRepository));
         }
 
         public async Task<ForumPost> CreateForumPost(ForumPost post)
@@ -36,12 +39,34 @@ namespace TheAgoraAPI.Repositories
 
         public async Task<int> DeleteForumPost(int id)
         {
+            var post = await dbContext.ForumPosts
+                .Include(p => p.ForumComments)
+                .ThenInclude(fc => fc.Comment) // Include the associated Comment
+                .FirstOrDefaultAsync(p => p.PostId == id);
+
+            if (post == null)
+            {
+                return 0;
+            }
+
+            // Collect the associated Comment entities
+            var commentsToDelete = post.ForumComments.Select(fc => fc.Comment).ToList();
+
+            // Remove associated forum comments first
+            dbContext.ForumComments.RemoveRange(post.ForumComments);
+
+            // Then, remove the comments
+            dbContext.Comments.RemoveRange(commentsToDelete);
+
             // First, remove all likes associated with this post
             await likeRepository.RemoveAllLikesForPostAsync(id);
 
             // Then, delete the forum post
             var rowsAffected = await dbContext.ForumPosts.Where(x => x.PostId == id).ExecuteDeleteAsync();
+
             return rowsAffected;
         }
+
+
     }
 }
